+++ CHANGELOG.md
@@ -9,9 +9,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixed
 ### Added
 
-## [Unreleased]
+## [100.1.8] - 2023-04-23
 ### Added
 - Asynchronous sending emails by cron
+- template configuration for admin confirmation email
+- set replyto on admin confiration email
 
 ## [100.1.7] - 2023-02-10
 ### Fixed